+++ src/events/instance.js
@@ -1,0 +1,3 @@
+import LocalEmitter from "../events/LocalEmitter";
+
+export default new LocalEmitter();